bugfix_20610

+++ Modules/Survey/classes/class.ilObjSurvey.php
@@ -1968,8 +1968,12 @@ class ilObjSurvey extends ilObject
 	{
 		include_once "./Modules/SurveyQuestionPool/classes/class.SurveyQuestion.php";
 		$question =& $this->_instanciateQuestion($question_id);
-		$question->delete($question_id);
-		$this->removeConstraintsConcerningQuestion($question_id);
+		#20610 if no question found, do nothing.
+		if($question)
+		{
+			$question->delete($question_id);
+			$this->removeConstraintsConcerningQuestion($question_id);
+		}
 	}
 	
 /**
